requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
Werkzeug==0.9.4
cmsplugin-filer==0.9.5
django-classy-tags==0.5.1
django-cms==3.0.1
django-debug-toolbar==1.2.1
django-extensions==1.3.7
django-filer==0.9.5
django-mptt==0.5.5
django-polymorphic==0.5.5
django-reversion==1.8.0
django-sekizai==0.7
django-session-security==2.2.0
djangocms-admin-style==0.2.2
djangocms-text-ckeditor==2.1.4
easy-thumbnails==2.0.1
html5lib==0.999
ipython==2.0.0
pyreadline==2.0
six==1.6.1
sqlparse==0.1.11
wsgiref==0.1.2
